Fix optional properties for pre-material systems (#18999)
