+++ framework/src/actions/ResolveOptionalMaterialPropertiesAction.C
@@ -8,8 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ResolveOptionalMaterialPropertiesAction.h"
-#include "Material.h"
-#include "FEProblem.h"
+#include "MaterialPropertyInterface.h"
 
 registerMooseAction("MooseApp",
                     ResolveOptionalMaterialPropertiesAction,
@@ -33,18 +32,7 @@ ResolveOptionalMaterialPropertiesAction::act()
 {
   mooseAssert(_problem, "Problem doesn't exist");
 
-  const Moose::MaterialDataType mdt_list[] = {Moose::BLOCK_MATERIAL_DATA,
-                                              Moose::FACE_MATERIAL_DATA,
-                                              Moose::NEIGHBOR_MATERIAL_DATA,
-                                              Moose::INTERFACE_MATERIAL_DATA};
-
-  for (auto mdt : mdt_list)
-  {
-    auto & materials = _problem->getMaterialWarehouse()[mdt];
-
-    for (auto tid : make_range(materials.numThreads()))
-      for (auto matbase_ptr : materials.getObjects(tid))
-        if (auto mat_ptr = std::dynamic_pointer_cast<Material>(matbase_ptr))
-          mat_ptr->resolveOptionalProperties();
-  }
+  const auto & mpi_registry = MaterialPropertyInterface::getInterfaceObjects(_app);
+  for (auto mpi : mpi_registry)
+    mpi->resolveOptionalProperties();
 }
