+++ framework/src/materials/Material.C
@@ -147,10 +147,3 @@ Material::computeConstantOption()
 
   return co;
 }
-
-void
-Material::resolveOptionalProperties()
-{
-  for (auto & proxy : _optional_property_proxies)
-    proxy->resolve(*this);
-}
