+++ framework/src/materials/MaterialPropertyInterface.C
@@ -15,6 +15,10 @@
 
 defineLegacyParams(MaterialPropertyInterface);
 
+// static material property interface object registry used for optional properties
+std::map<MooseApp *, std::vector<MaterialPropertyInterface *>>
+    MaterialPropertyInterface::_mpi_registry;
+
 InputParameters
 MaterialPropertyInterface::validParams()
 {
@@ -46,7 +50,6 @@ MaterialPropertyInterface::MaterialPropertyInterface(const MooseObject * moose_o
     _mi_block_ids(block_ids),
     _mi_boundary_ids(boundary_ids)
 {
-
   // Set the MaterialDataType flag
   if (_mi_params.isParamValid("_material_data_type"))
     _material_data_type = _mi_params.get<Moose::MaterialDataType>("_material_data_type");
@@ -59,6 +62,9 @@ MaterialPropertyInterface::MaterialPropertyInterface(const MooseObject * moose_o
 
   _material_data =
       _mi_feproblem.getMaterialData(_material_data_type, _mi_params.get<THREAD_ID>("_tid"));
+
+  // register self
+  _mpi_registry[&moose_object->getMooseApp()].push_back(this);
 }
 
 std::string
@@ -316,4 +322,19 @@ MaterialPropertyInterface::checkExecutionStage()
   if (_mi_feproblem.startedInitialSetup())
     mooseError("Material properties must be retrieved during object construction. This is a code "
                "problem.");
+}
+
+const std::vector<MaterialPropertyInterface *>
+MaterialPropertyInterface::getInterfaceObjects(MooseApp & app)
+{
+  // insertion is intentional here to return a reference to an empty vector even
+  // if no objects derived from material property interface are used in the model
+  return _mpi_registry[&app];
+}
+
+void
+MaterialPropertyInterface::resolveOptionalProperties()
+{
+  for (auto & proxy : _optional_property_proxies)
+    proxy->resolve(*this);
 }
